Added websocket for course chat
feat:added websocket for course chat

+++ mentorlib/apps/courses/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect
+from django.contrib.auth.decorators import login_required
+
 from mentorlib.apps.courses.models import (
     Course,
     AskedCourse,
     CourseRegisteredStudent,
     CourseUploadFile,
+    Comment,
 )
 from mentorlib.apps.users.models import User, UserUpload
 from mentorlib.apps.configuration.models import Resource
@@ -18,12 +21,13 @@ from pathlib import Path
 
 NAV_URLS = {
     "course_nav": {
-        "Informations":"courses:course_details",
-        "Etudiant inscrits":"courses:course_students_registered",
-        "Documents":"courses:course_files"
+        "Informations": "courses:course_details",
+        "Etudiant inscrits": "courses:course_students_registered",
+        "Documents": "courses:course_files",
     },
 }
-CONTEXT = { **NAV_URLS, 'messages':[]}
+CONTEXT = {**NAV_URLS, "messages": []}
+
 
 def index(request):
     CONTEXT["courses"] = CourseFilter(request.GET, queryset=Course.objects.all())
@@ -32,44 +36,47 @@ def index(request):
 
 
 def course_details(request, id):
+    """Show basic course informations and handle comments messages"""
     course = Course.objects.get(id=id)
-    CONTEXT['course'] = course
-    CONTEXT['file_upload'] = CourseUploadForm()
+    CONTEXT["course"] = course
 
     action = request.GET.get("action")
 
-    if action:
-        course_registered = CourseRegisteredStudent.objects.filter(
-            student=request.user, course=course
-        ).first()
-        if action == "unsuscribe":
-            if course_registered:
-                course_registered.delete()
-                CONTEXT["messages"].append(
-                    {"color": "red", "text": "Vous vous êtes désinscrit"}
-                )
-            else:
-                CONTEXT["messages"].append(
-                    {"color": "red", "text": "Vous n'êtes pas inscrit"}
-                )
-        elif action == "suscribe" and not course_registered:
-            course_registered = CourseRegisteredStudent(
+    if not request.user.is_anonymous:
+        if action:
+            course_registered = CourseRegisteredStudent.objects.filter(
                 student=request.user, course=course
-            )
-            course_registered.save()
-            CONTEXT["messages"].append(
-                {
-                    "color": "green",
-                    "text": "Votre inscription a bien été prit en compte",
-                }
+            ).first()
+            if action == "unsuscribe":
+                if course_registered:
+                    course_registered.delete()
+                    CONTEXT["messages"].append(
+                        {"color": "red", "text": "Vous vous êtes désinscrit"}
+                    )
+                else:
+                    CONTEXT["messages"].append(
+                        {"color": "red", "text": "Vous n'êtes pas inscrit"}
+                    )
+            elif action == "suscribe" and not course_registered:
+                course_registered = CourseRegisteredStudent(
+                    student=request.user, course=course
+                )
+                course_registered.save()
+                CONTEXT["messages"].append(
+                    {
+                        "color": "green",
+                        "text": "Votre inscription a bien été prit en compte",
+                    }
+                )
+
+        if request.method == "POST" and request.POST.get("message"):
+            comment = Comment(
+                course=Course.objects.get(id=id),
+                comment=request.POST.get("message"),
+                user=request.user,
             )
 
-    if request.method == "POST" and request.POST.get("student_note"):
-        course_registered = CourseRegisteredStudent.objects.get(
-            student=request.user, course=course
-        )
-        course_registered.note = request.POST.get("student_note")
-        course_registered.save()
+            comment.save()
 
     return render(request, "courses/tabs/informations.html", context=CONTEXT)
 
@@ -79,11 +86,9 @@ def duration_to_minutes(duration: str):
     return int(duration_split[0]) * 60 + int(duration_split[1])
 
 
+@login_required(redirect_field_name="my_redirect_field")
 @requires_csrf_token
 def course_create(request):
-    if not request.user.is_authenticated:
-        return redirect("/users/login")
-
     if request.method == "POST":
         convert_date = datetime.strptime(
             f"{request.POST['date']} {request.POST['time']}", "%m/%d/%Y %H:%M"
@@ -157,5 +162,12 @@ def course_views_file(request, id, file_id):
 def course_students_registered(request, id):
     course = Course.objects.get(id=id)
     CONTEXT["course"] = course
+
+    if request.method == "POST" and request.POST.get("student_note"):
+        course_registered = CourseRegisteredStudent.objects.get(
+            student=request.user, course=course
+        )
+        course_registered.note = request.POST.get("student_note")
+        course_registered.save()
 
     return render(request, "courses/tabs/register_informations.html", context=CONTEXT)
